Fix scss linting
The configuration file was being ignored due to bugs and incorrect
settings requiring some modifications to the globbing but primarily an
upgrade to the latest released version. Turns out `scss-lint` was
deprecated in favour of `scss_lint` so we needed to switch over to get
recent updates (since 0.3x)

The new linter introduced the `PseudoElement` rule that raises some
issues in `_guider-appointments.scss` that I resolved.