Fix orphaned bookable slots
There was a really irritating issue where sometimes we just couldn't
book appointments, randomly. This was because the bookable slot that
matched didn't have an existing user anymore, because we deleted that
user in db:seed.

The fix is to ensure we delete any child bookable slots when we delete
users.