Batch updates on appointments
Allows the user to 'commit' a bunch of changes made to multiple
appointments via the resource calendar.

Error handling has been neatly side-stepped for now but probably the
best thing to do is process each and at least give some notification
that a single one failed if any. I'll revisit this.