Allow any user to book and reschedule appointments
Sometimes a guider will call the customer at the scheduled time,
and the customer is no longer available. Often the guider will
simply book another appointment for the customer at this stage.

There are other times that TPAS will want to book appointments in
the case that the customer has called through via a means other
than the contact centre.