Notify upon appointment changes
This hooks up pusher for real-time notifications of appointment
changes, namely reassignments to different guiders. Both the original
guider and the newly assigned guider are notified when necessary.
Otherwise just the already specified guider is notified when the
appointment time / date / duration changes.

This starts a fake pusher server in a separate process so we can go
fully end-to-end in the spec.

The spec has to jump back and forth between user sessions, but I've
commented to hopefully give a steer.