Display only my appointments
When the user is both a resource manager and a guider - the scoping
logic for the appointments JSON would short-circuit to all appointments
due to ordering of conditionals favouring the resource manager
permission.

This change ensures that when the user is both guider and manager they
only see appointments scoped directly to them - i.e. appointments they
are holding with the customer.