Import Booking Bug appointments
This imports the existing booking bug appointments from the provided
CSV URI or file path.

Some things to note:

- The guider UIDs need to map correctly to staging / prod signon
- Appointments will be booked by the 'Pension Wise Import' agent
- If the guider isn't mapped they're marked `disabled`
- It's kinda slow but that's OK - it's an ad-hoc thing

Appointment creation is not validated since doing so would cause
problems with the two working day validations. As it stands we can be
pretty certain that the data from the CSV is good.

The specs are intentionally light since the nature of this import is
that it will need eyes on the imported data.