Refine allocation / assignment messages
The way these messages were worded prior to this change implied the
assignment was a user-driven activity. This is incorrect since
allocation is randomised and assigned via the system.

In the case of reassignment we now display richer text describing which
resource manager made the selection and which guider was selected.