Add missing index on `start_at`
The recent `#stale?` addition performs checks to determine the freshness
of the given model / model collection instance by counting the elements
and retrieving the `MAX("appointments"."updated_at")`, merging the
same `where` clauses for the given model / model collection.

Since we had no index on this column this query was very slow (at least
in postgresql terms) in the order of `~100ms`. Adding this index drops
this down to `~1ms` which is quite some gain.