Parallelise specs across multiple processes
Run specs in parallel across your multiple cores / processes. On my
machine this drops the spec runs to around 35 seconds.

You can still run `rake` as normal or `rspec` if you want to run specs
serially so this is opt-in for now. To get this setup on your machine
first pull this branch, then:

        $ bundle exec rake parallel:create

This creates multiple databases to run against.

        $ bundle exec rake parallel:prepare

This ensures each database is at the same version, migration-wise. You
should run this after running migrations as normal too.

Run the specs:

        $ bundle exec parallel_rspec

You can choose how many processes to run against by exporting
`PARALLEL_TEST_PROCESSORS=X` where `X` is the number of processes. A
good rule of thumb is to use the same number as physical cores.