Fix bug where all day holidays were showing up as 1 day less in length
FullCalendar shows all day events exclusive of the end date, so
that means if we have a two day holiday, it will always show up as one
day in the all day slot.
To get around this we add one day to all holidays that are "all day".
Also see: http://stackoverflow.com/questions/27604359/fullcalendar-event-spanning-all-day-are-one-day-too-short

Also, validate Holiday start_at and end_at