Quick and dirty script to fix current records
Given a CSV with the correct mappings from member ID to booking
reference this will find the existing appointments and replace their
IDs along with any associations.

This is a one hit and quit affair that will be removed after it has
run. I've given this a thorough work out locally and all is well
against a copy of the production system.