Add bookable slots API for customer booking
This adds a new JSON API endpoint at `/api/v1/bookable_slots` and
returns a payload like:

```
[
  "2017-01-01": [ "12:00", "13:00", "15:00" ],
  "2017-01-02": [ "12:10", "14:10" ]
]
```

It requires the `pension_wise_api` permission to be present for the
requesting user.

This first cut relies on grouping, de-duping and sorting in memory which
isn't ideal. The alternative turned into the beastiest SQL query I've
ever written and wasn't ideal for maintenance purposes, though a great
deal more performant.

Running this across the booking window (6 weeks) on a copy of the
production database, but in development mode yields the full response
in ~700ms.