Introduce bookings API
This provides the basis for the bookings API at:

`POST /api/v1/appointments`

Callers will need the `pension_wise_api` permission from signon.

The form object is versioned using the same scheme as the API but since
the model and form classes are the same and Rails scoping is stupid,
I've had to dud alias the model class under a `Models` namespace so we
can qualify the scoping when calling the model `Appointment` from the
API `Appointment`.