Validate appointment API requests
Ensures appointment payloads are valid before attempting to book
appointments. The email is required when created by the API so we
explicitly check for this.

When determined invalid we return a 422 and the validation errors
serialized as:

```
[
  "attribute_name": [
    "an error description",
    "another description"
  ],
  "another_attribute_name": [
    "another description"
  ]
]
```

The presence of `guider` in the errors serves as a special case for an
already booked slot. In practice this is absolutely an edgecase but
could potentially occur in an (albeit unlikely) race condition.