Limit cartesian products on join filters
Given the size and cartesian products of the various join filters
required to infer and limit the bookable slots, we were reducing down
from multiple millions of rows. Limiting the scope and range of the
join filters speeds this up from multiple seconds (across our working
set) to just a few milliseconds.