Allow for appointments being displayed in UTC
This is a temporary hack, please forgive me.

Appointments are stored / displayed in UTC but we're currently in BST.
An appointment scheduled for '12:20' actually takes place at '11:20'
according to the server, so when the appointment is updated after it
ends the `future?` check passes since '12:20' is greater than '11:20'.
If we substract an hour from the appointment's start time, we end up
with a true representation of time to compare with.