Update appointment type from date of birth
This change preselects the correct appointment type from the provided
date of birth when booking / modifying an appointment. It's possible the
agent, guider or resource manager will want to override this if they
spot the appointment is within a few months of a particular range - so
this is also possible.