Ensure notifications are directed accurately
In the case of dropped summary documents these are always the guider's
responsibility to action, so only they will receive the high-priority
notifications.

In the case of other dropped emails - for now we will notify both the
guider and the agent - so long as the agent is not the Pension Wise
website API agent. This is despite the `owner` of the activity always
being the guider.

Since these two activities are always considered 'high priority' we can
do away with the guard and move the event hooks into the concrete
activity classes.