Push the email guard into the job itself
This moves the guard as low as possible, so we don't get caught out
when calling this job with an appointment not satisfying the criterion
for email notification - namely an appointment without a
customer-supplied email address.

I took an opportunity to simplify the specs slightly by making them
check state changes rather than the interaction between the various
stubs.