Create recurring holidays
When creating single day holidays with an associated start and end
time, the resource manager now has the option to specify recurrence.
When they provide a recurrence date the holiday is repeated for each
weekday between the start date and the provided end date. This behaves
similarly to all-day holidays but with the option to have shorter
periods for each day.