Allow resource managers to adhoc reschedule
Grants resource managers the ability to reschedule appointments on an
adhoc basis, ie not from the calendar schedule. When rescheduling the
resource manager is presented with a further option to select from
adhoc availability. When they select this they are then presented with
a date and guider to choose from.