Bump gem dependencies
Bumps gem dependencies with a simple `bundle update`.

I had to change the default for `permissions` on the `User` model since
we use JSONB and the GDS-SSO linter now tries to set `permissions` to a
null value during testing. We were previously not allowing nulls on
this column. It has no real side-effects for our usage anyway.