Summarise availability data for reporting
This will be scheduled to run every morning before the day's
appointments are due to start. It summarises availability over the
following four week period, inside of the regular booking window.

At the moment this is only concerned with TPAS guider appointments but
once we support multiple providers it will address other provider's
appointments also.