Improve layout of navigation bar
This would wrap for users with the administrator role, since the
provider dropdown was styled incorrectly and has recently increased in
width. The app title has changed to drop the 'Appointment', this matches
the other planners that also changed recently and has the useful effect
of freeing up some valuable horizontal real estate.