Third party bookings
Allows agents and others to specify a third party will take the
appointment on behalf of the given customer. When this is specified, the
agent will also be asked to confirm the customer gave their consent.
They can request a printed consent form via notify, or specify power of
attorney. In both instances, they can upload evidence of both.

Resource managers will receive email notification when the third party
option is specified. This is now combined with what was previously the
accessibility adjustments email. A banner will also be displayed when
editing/viewing the appointment.