Begin switch to webdrivers
PhantomJS has been dead for a long time and for various reasons we were
still pinned to old versions of the supporting testing libraries. This
change begins the migration towards chrome-driver instead. Some of the
specs are still tagged with the poltergeist driver, these can be
resolved as time goes by.