Improve searching
This is a performance nightmare as it currently stands. We can
short-circuit a lot of queries straight to a simple where clause when
the query is just an appointment reference. Also, we can bound the
result set when no date range is specified, so we're not performing text
scans across the whole dataset.