Determine correct grace periods
When the user is a TPAS agent, we were in certain cases still presenting
them with no grace period when returning slots. There were further
complications when a booking is rescheduled. To counter these, we have
to pass along both the user in context and the contextual guider to the
given appointment, in the case of rescheduling specifically.

Outside of TPAS agents, everything remains the same.