Load chromedriver robustly
This has been causing issues for a while and was compounded by travis CI
installing the latest chrome but not the latest chromedriver.

We can use selenium-webdriver to resolve the chromedriver version in
Ruby.

I've had to temporarily disable some tests that are failing due to a
change in how coordinates are resolved in capybara/selenium. They appear
to be relative to the middle of the element rather than the top left as
previous.