Ensure allocation respects booking window
Prior to this change the allocation of slots was not respecting the
variable booking window per provider. For example, in the case of
TPAS/ops, they could have 10 slots for the given `start_at` and this
would outweigh the slot that would usually be allocated by a single
instance of matching date/time despite being outside of their regular
booking window.