Remove guider's availability when SSO suspended
When a guider is suspended or has the 'guider' permission removed in
signon, there are some cases where their availability continues and
errant appointments are created with their slots. This change ensures
that when a guider is suspended, or has their 'guider' permission
removed, their account is marked inactive and their future availability
is removed.

When they are reinstated, their resource manager can create new
availability schedules as they see fit.