Find appointments with normalised numbers
There have been incidents where customers replied to their reminder SMS
messages with the 'cancel' keyword but received a cancellation failure
SMS in response. These are due to normalisation issues occurring between
us and Notify. For example, when a customer's number is stored as
'+447...' the `source_number` we receive back from Notify is `447...'
thus the appointment for cancellation cannot be found. This change will
attempt to find the number by many sensible variants instead.