Reset `casebook_appointment_id` after cancellation
There is no reason to retain this after an appointment is successfully
cancelled via casebook. It causes issues further down the lifecycle of
an appointment when the appointment was casebook cancelled (usually due
to being reallocated away from the originally owning casebook enrolled
guider to a non-enrolled guider) and causes multiple casebook
cancellation calls thereafter.

Cancellations on the casebook side appear to be idempotent so it doesn't
really cause a problem there, though it does result in multiple
cancellation entries in the appointment's activity feed, and so is
confusing to anyone viewing.