Ensure redaction occurs
There were cases where the appointment was in an 'invalid' state so
calling `#update` would fail silently. For example - when the
appointment had a primary status assigned that would require a secondary
status. Using `update_all` ensures the validations are skipped and the
updates are forced through.

We also now redact the `adjustments` attribute and any received SMS
messages in the activity feed.